+++ src/backend/api/routes/index.js
@@ -1,7 +1,9 @@
 import filesRoutes from "./files";
 import foldersRoutes from "./folders";
+import thumbnailsRoutes from "./thumbnails";
 
 export default (router) => {
   filesRoutes(router);
   foldersRoutes(router);
+  thumbnailsRoutes(router);
 };