+++ src/constants/routes.js
@@ -1,7 +1,10 @@
 export const API = "/api/v1";
 export const FILE_UPLOAD = `${API}/files/upload`;
 export const FILE_DOWNLOAD = `${API}/files/:fileId/download`;
+
+export const THUMBNAIL_UPLOAD = `${API}/thumbnail/upload`;
+
 export const FOLDER_LIST_ITEMS = `${API}/folders/:folderId/items`;
 export const FOLDER_CREATE = `${API}/folders/create`;
 export const FOLDER_THUMBNAILS = `${API}/folders/:folderId/thumbnails`;
-export const FOLDER_FIND = `${API}/folders/:folderPath/find`;
+export const FOLDER_FETCH = `${API}/folders/:folderPath/fetch`;